source/requirements.txt
# Job-Automatisierungs-Pipeline – Andreas Eichmann
# pip install -r requirements.txt

# Scraping
requests>=2.31
beautifulsoup4>=4.12
feedparser>=6.0
lxml>=5.0

# Dokumente
python-docx>=1.1
docx2pdf>=0.1.8

# Microsoft OAuth fuer Outlook SMTP
msal>=1.30

# Formular-Automatisierung (optional)
selenium>=4.20
webdriver-manager>=4.0
